requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
pydantic==2.10.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
gspread==6.1.4
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

